+++ drivers/rqlite/rqlite.go
@@ -164,14 +164,17 @@ func (d *driveri) Open(ctx context.Context, src *source.Source, _ driver.AccessM
 		return nil, err
 	}
 
-	if err = driver.OpeningPing(ctx, src, db); err != nil {
+	ver, err := driver.OpeningPing(ctx, src, db, d.DBSemver)
+	if err != nil {
 		return nil, enrichConnError(err, src)
 	}
 
-	return &grip{
+	g := &grip{
 		log: d.log, db: db, src: src, drvr: d,
 		drvrw: &enrichingSQLDriver{driveri: d, src: src},
-	}, nil
+	}
+	g.semver.Prime(ver)
+	return g, nil
 }
 
 func (d *driveri) doOpen(ctx context.Context, src *source.Source) (*sql.DB, error) {